+++ api/lib/infrastructure/jobs/campaign-result/SendSharedParticipationResultsToPoleEmploiHandler.js
@@ -1,4 +1,4 @@
-import { usecases } from '../../../domain/usecases/index.js';
+import { usecases } from '../../../../src/prescription/campaign-participation/domain/usecases/index.js';
 import { SendSharedParticipationResultsToPoleEmploiJob } from './SendSharedParticipationResultsToPoleEmploiJob.js';
 
 class SendSharedParticipationResultsToPoleEmploiHandler {
