+++ api/src/prescription/campaign-participation/domain/events/handle-pole-emploi-participation-finished.js
@@ -1,16 +1,17 @@
-import { PoleEmploiPayload } from '../../infrastructure/externals/pole-emploi/PoleEmploiPayload.js';
-import * as httpErrorsHelper from '../../infrastructure/http/errors-helper.js';
-import { httpAgent } from '../../infrastructure/http/http-agent.js';
-import * as monitoringTools from '../../infrastructure/monitoring-tools.js';
+import { AssessmentCompleted } from '../../../../../lib/domain/events/AssessmentCompleted.js';
+import { checkEventTypes } from '../../../../../lib/domain/events/check-event-types.js';
+import * as httpErrorsHelper from '../../../../../lib/infrastructure/http/errors-helper.js';
+import { httpAgent } from '../../../../../lib/infrastructure/http/http-agent.js';
+import { monitoringTools } from '../../../../../lib/infrastructure/monitoring-tools.js';
+import { PoleEmploiPayload } from '../../infrastructure/externals/PoleEmploiPayload.js';
 import { PoleEmploiSending } from '../models/PoleEmploiSending.js';
-import { CampaignParticipationStarted } from './CampaignParticipationStarted.js';
-import { checkEventTypes } from './check-event-types.js';
 
-const eventTypes = [CampaignParticipationStarted];
+const eventTypes = [AssessmentCompleted];
 
-async function handlePoleEmploiParticipationStarted({
+async function handlePoleEmploiParticipationFinished({
   event,
   authenticationMethodRepository,
+  assessmentRepository,
   campaignRepository,
   campaignParticipationRepository,
   organizationRepository,
@@ -23,6 +24,8 @@ async function handlePoleEmploiParticipationStarted({
 
   const { campaignParticipationId } = event;
 
+  if (!campaignParticipationId) return;
+
   const participation = await campaignParticipationRepository.get(campaignParticipationId);
   const campaign = await campaignRepository.get(participation.campaignId);
   const organization = await organizationRepository.get(campaign.organizationId);
@@ -30,14 +33,15 @@ async function handlePoleEmploiParticipationStarted({
   if (campaign.isAssessment() && organization.isPoleEmploi) {
     const user = await userRepository.get(participation.userId);
     const targetProfile = await targetProfileRepository.get(campaign.targetProfileId);
+    const assessment = await assessmentRepository.get(participation.lastAssessment.id);
 
-    const payload = PoleEmploiPayload.buildForParticipationStarted({
+    const payload = PoleEmploiPayload.buildForParticipationFinished({
       user,
       campaign,
       targetProfile,
       participation,
+      assessment,
     });
-
     const response = await poleEmploiNotifier.notify(user.id, payload, {
       authenticationMethodRepository,
       httpAgent,
@@ -45,7 +49,7 @@ async function handlePoleEmploiParticipationStarted({
       monitoringTools,
     });
 
-    const poleEmploiSending = PoleEmploiSending.buildForParticipationStarted({
+    const poleEmploiSending = PoleEmploiSending.buildForParticipationFinished({
       campaignParticipationId,
       payload: payload.toString(),
       isSuccessful: response.isSuccessful,
@@ -56,5 +60,5 @@ async function handlePoleEmploiParticipationStarted({
   }
 }
 
-handlePoleEmploiParticipationStarted.eventTypes = eventTypes;
-export { handlePoleEmploiParticipationStarted };
+handlePoleEmploiParticipationFinished.eventTypes = eventTypes;
+export { handlePoleEmploiParticipationFinished };
