+++ api/src/prescription/campaign-participation/domain/events/handle-pole-emploi-participation-started.js
@@ -1,17 +1,15 @@
-import { PoleEmploiPayload } from '../../infrastructure/externals/pole-emploi/PoleEmploiPayload.js';
-import * as httpErrorsHelper from '../../infrastructure/http/errors-helper.js';
-import { httpAgent } from '../../infrastructure/http/http-agent.js';
-import * as monitoringTools from '../../infrastructure/monitoring-tools.js';
+import { CampaignParticipationStarted } from '../../../../../lib/domain/events/CampaignParticipationStarted.js';
+import { checkEventTypes } from '../../../../../lib/domain/events/check-event-types.js';
+import * as httpErrorsHelper from '../../../../../lib/infrastructure/http/errors-helper.js';
+import { httpAgent } from '../../../../../lib/infrastructure/http/http-agent.js';
+import { monitoringTools } from '../../../../../lib/infrastructure/monitoring-tools.js';
+import { PoleEmploiPayload } from '../../infrastructure/externals/PoleEmploiPayload.js';
 import { PoleEmploiSending } from '../models/PoleEmploiSending.js';
-import { AssessmentCompleted } from './AssessmentCompleted.js';
-import { checkEventTypes } from './check-event-types.js';
+const eventTypes = [CampaignParticipationStarted];
 
-const eventTypes = [AssessmentCompleted];
-
-async function handlePoleEmploiParticipationFinished({
+async function handlePoleEmploiParticipationStarted({
   event,
   authenticationMethodRepository,
-  assessmentRepository,
   campaignRepository,
   campaignParticipationRepository,
   organizationRepository,
@@ -24,8 +22,6 @@ async function handlePoleEmploiParticipationFinished({
 
   const { campaignParticipationId } = event;
 
-  if (!campaignParticipationId) return;
-
   const participation = await campaignParticipationRepository.get(campaignParticipationId);
   const campaign = await campaignRepository.get(participation.campaignId);
   const organization = await organizationRepository.get(campaign.organizationId);
@@ -33,15 +29,14 @@ async function handlePoleEmploiParticipationFinished({
   if (campaign.isAssessment() && organization.isPoleEmploi) {
     const user = await userRepository.get(participation.userId);
     const targetProfile = await targetProfileRepository.get(campaign.targetProfileId);
-    const assessment = await assessmentRepository.get(participation.lastAssessment.id);
 
-    const payload = PoleEmploiPayload.buildForParticipationFinished({
+    const payload = PoleEmploiPayload.buildForParticipationStarted({
       user,
       campaign,
       targetProfile,
       participation,
-      assessment,
     });
+
     const response = await poleEmploiNotifier.notify(user.id, payload, {
       authenticationMethodRepository,
       httpAgent,
@@ -49,7 +44,7 @@ async function handlePoleEmploiParticipationFinished({
       monitoringTools,
     });
 
-    const poleEmploiSending = PoleEmploiSending.buildForParticipationFinished({
+    const poleEmploiSending = PoleEmploiSending.buildForParticipationStarted({
       campaignParticipationId,
       payload: payload.toString(),
       isSuccessful: response.isSuccessful,
@@ -60,5 +55,5 @@ async function handlePoleEmploiParticipationFinished({
   }
 }
 
-handlePoleEmploiParticipationFinished.eventTypes = eventTypes;
-export { handlePoleEmploiParticipationFinished };
+handlePoleEmploiParticipationStarted.eventTypes = eventTypes;
+export { handlePoleEmploiParticipationStarted };
