+++ api/src/prescription/campaign-participation/domain/services/pole-emploi-service.js
@@ -1,4 +1,4 @@
-import { config } from '../../config.js';
+import { config } from '../../../../../lib/config.js';
 
 function generateLink(sending, filters = {}) {
   const host = config.apiManager.url;
