+++ api/src/prescription/campaign-participation/infrastructure/externals/pole-emploi-notifier.js
@@ -6,10 +6,10 @@ import querystring from 'node:querystring';
 
 import dayjs from 'dayjs';
 
-import * as OidcIdentityProviders from '../../../../src/identity-access-management/domain/constants/oidc-identity-providers.js';
-import { AuthenticationMethod } from '../../../../src/identity-access-management/domain/models/AuthenticationMethod.js';
-import { config } from '../../../config.js';
-import { UnexpectedUserAccountError } from '../../../domain/errors.js';
+import { UnexpectedUserAccountError } from '../../../../../lib/domain/errors.js';
+import * as OidcIdentityProviders from '../../../../identity-access-management/domain/constants/oidc-identity-providers.js';
+import { AuthenticationMethod } from '../../../../identity-access-management/domain/models/AuthenticationMethod.js';
+import { config } from '../../../../shared/config.js';
 
 const notify = async (userId, payload, dependencies) => {
   const { authenticationMethodRepository, httpAgent, httpErrorsHelper, monitoringTools } = dependencies;
