+++ api/src/prescription/campaign-participation/infrastructure/open-api-doc/pole-emploi-sendings-doc.js
