+++ api/src/shared/infrastructure/open-api-doc/erreur-doc.js
