+++ api/tests/prescription/campaign-participation/integration/domain/events/handle-pole-emploi-participation-finished_test.js
@@ -1,12 +1,12 @@
-import { CampaignParticipationStarted as CampaignParticipationStartedEvent } from '../../../../lib/domain/events/CampaignParticipationStarted.js';
-import { handlePoleEmploiParticipationStarted } from '../../../../lib/domain/events/handle-pole-emploi-participation-started.js';
-import * as campaignParticipationRepository from '../../../../lib/infrastructure/repositories/campaign-participation-repository.js';
-import * as campaignRepository from '../../../../lib/infrastructure/repositories/campaign-repository.js';
-import * as poleEmploiSendingRepository from '../../../../lib/infrastructure/repositories/pole-emploi-sending-repository.js';
-import * as targetProfileRepository from '../../../../lib/infrastructure/repositories/target-profile-repository.js';
-import * as assessmentRepository from '../../../../src/shared/infrastructure/repositories/assessment-repository.js';
-import * as organizationRepository from '../../../../src/shared/infrastructure/repositories/organization-repository.js';
-import * as userRepository from '../../../../src/shared/infrastructure/repositories/user-repository.js';
+import { AssessmentCompleted as AsessmentCompletedEvent } from '../../../../../../lib/domain/events/AssessmentCompleted.js';
+import * as campaignParticipationRepository from '../../../../../../lib/infrastructure/repositories/campaign-participation-repository.js';
+import * as campaignRepository from '../../../../../../lib/infrastructure/repositories/campaign-repository.js';
+import * as targetProfileRepository from '../../../../../../lib/infrastructure/repositories/target-profile-repository.js';
+import { handlePoleEmploiParticipationFinished } from '../../../../../../src/prescription/campaign-participation/domain/events/handle-pole-emploi-participation-finished.js';
+import * as poleEmploiSendingRepository from '../../../../../../src/prescription/campaign-participation/infrastructure/repositories/pole-emploi-sending-repository.js';
+import * as assessmentRepository from '../../../../../../src/shared/infrastructure/repositories/assessment-repository.js';
+import * as organizationRepository from '../../../../../../src/shared/infrastructure/repositories/organization-repository.js';
+import * as userRepository from '../../../../../../src/shared/infrastructure/repositories/user-repository.js';
 import {
   databaseBuilder,
   expect,
@@ -14,12 +14,12 @@ import {
   learningContentBuilder,
   mockLearningContent,
   sinon,
-} from '../../../test-helper.js';
+} from '../../../../../test-helper.js';
 
-describe('Integration | Event | Handle Pole emploi participation started', function () {
+describe('Integration | Event | Handle Pole emploi participation finished', function () {
   let campaignParticipationId, userId, event, poleEmploiNotifier, responseCode;
 
-  describe('#handlePoleEmploiParticipationStarted', function () {
+  describe('#handlePoleEmploiParticipationFinished', function () {
     beforeEach(async function () {
       responseCode = Symbol('responseCode');
       poleEmploiNotifier = { notify: sinon.stub().resolves({ isSuccessful: true, code: responseCode }) };
@@ -34,7 +34,7 @@ describe('Integration | Event | Handle Pole emploi participation started', funct
       const campaignId = databaseBuilder.factory.buildCampaign({ targetProfileId, organizationId }).id;
       campaignParticipationId = databaseBuilder.factory.buildCampaignParticipation({ campaignId, userId }).id;
       databaseBuilder.factory.buildAssessment({ campaignParticipationId, userId });
-      event = new CampaignParticipationStartedEvent();
+      event = new AsessmentCompletedEvent();
       event.campaignParticipationId = campaignParticipationId;
 
       const learningContentObjects = learningContentBuilder.fromAreas([]);
@@ -45,7 +45,7 @@ describe('Integration | Event | Handle Pole emploi participation started', funct
 
     it('should notify pole emploi and save success of this notification', async function () {
       // when
-      await handlePoleEmploiParticipationStarted({
+      await handlePoleEmploiParticipationFinished({
         event,
         assessmentRepository,
         campaignRepository,
@@ -61,7 +61,7 @@ describe('Integration | Event | Handle Pole emploi participation started', funct
       const poleEmploiSendings = await knex('pole-emploi-sendings').where({ campaignParticipationId });
       expect(poleEmploiSendings.length).to.equal(1);
       expect(poleEmploiSendings[0].responseCode).to.equal(responseCode.toString());
-      expect(poleEmploiSendings[0].type).to.equal('CAMPAIGN_PARTICIPATION_START');
+      expect(poleEmploiSendings[0].type).to.equal('CAMPAIGN_PARTICIPATION_COMPLETION');
     });
   });
 });
