+++ api/tests/prescription/campaign-participation/integration/domain/events/handle-pole-emploi-participation-started_test.js
@@ -1,12 +1,12 @@
-import { AssessmentCompleted as AsessmentCompletedEvent } from '../../../../lib/domain/events/AssessmentCompleted.js';
-import { handlePoleEmploiParticipationFinished } from '../../../../lib/domain/events/handle-pole-emploi-participation-finished.js';
-import * as campaignParticipationRepository from '../../../../lib/infrastructure/repositories/campaign-participation-repository.js';
-import * as campaignRepository from '../../../../lib/infrastructure/repositories/campaign-repository.js';
-import * as poleEmploiSendingRepository from '../../../../lib/infrastructure/repositories/pole-emploi-sending-repository.js';
-import * as targetProfileRepository from '../../../../lib/infrastructure/repositories/target-profile-repository.js';
-import * as assessmentRepository from '../../../../src/shared/infrastructure/repositories/assessment-repository.js';
-import * as organizationRepository from '../../../../src/shared/infrastructure/repositories/organization-repository.js';
-import * as userRepository from '../../../../src/shared/infrastructure/repositories/user-repository.js';
+import { CampaignParticipationStarted as CampaignParticipationStartedEvent } from '../../../../../../lib/domain/events/CampaignParticipationStarted.js';
+import * as campaignParticipationRepository from '../../../../../../lib/infrastructure/repositories/campaign-participation-repository.js';
+import * as campaignRepository from '../../../../../../lib/infrastructure/repositories/campaign-repository.js';
+import * as targetProfileRepository from '../../../../../../lib/infrastructure/repositories/target-profile-repository.js';
+import { handlePoleEmploiParticipationStarted } from '../../../../../../src/prescription/campaign-participation/domain/events/handle-pole-emploi-participation-started.js';
+import * as poleEmploiSendingRepository from '../../../../../../src/prescription/campaign-participation/infrastructure/repositories/pole-emploi-sending-repository.js';
+import * as assessmentRepository from '../../../../../../src/shared/infrastructure/repositories/assessment-repository.js';
+import * as organizationRepository from '../../../../../../src/shared/infrastructure/repositories/organization-repository.js';
+import * as userRepository from '../../../../../../src/shared/infrastructure/repositories/user-repository.js';
 import {
   databaseBuilder,
   expect,
@@ -14,12 +14,12 @@ import {
   learningContentBuilder,
   mockLearningContent,
   sinon,
-} from '../../../test-helper.js';
+} from '../../../../../test-helper.js';
 
-describe('Integration | Event | Handle Pole emploi participation finished', function () {
+describe('Integration | Event | Handle Pole emploi participation started', function () {
   let campaignParticipationId, userId, event, poleEmploiNotifier, responseCode;
 
-  describe('#handlePoleEmploiParticipationFinished', function () {
+  describe('#handlePoleEmploiParticipationStarted', function () {
     beforeEach(async function () {
       responseCode = Symbol('responseCode');
       poleEmploiNotifier = { notify: sinon.stub().resolves({ isSuccessful: true, code: responseCode }) };
@@ -34,7 +34,7 @@ describe('Integration | Event | Handle Pole emploi participation finished', func
       const campaignId = databaseBuilder.factory.buildCampaign({ targetProfileId, organizationId }).id;
       campaignParticipationId = databaseBuilder.factory.buildCampaignParticipation({ campaignId, userId }).id;
       databaseBuilder.factory.buildAssessment({ campaignParticipationId, userId });
-      event = new AsessmentCompletedEvent();
+      event = new CampaignParticipationStartedEvent();
       event.campaignParticipationId = campaignParticipationId;
 
       const learningContentObjects = learningContentBuilder.fromAreas([]);
@@ -45,7 +45,7 @@ describe('Integration | Event | Handle Pole emploi participation finished', func
 
     it('should notify pole emploi and save success of this notification', async function () {
       // when
-      await handlePoleEmploiParticipationFinished({
+      await handlePoleEmploiParticipationStarted({
         event,
         assessmentRepository,
         campaignRepository,
@@ -61,7 +61,7 @@ describe('Integration | Event | Handle Pole emploi participation finished', func
       const poleEmploiSendings = await knex('pole-emploi-sendings').where({ campaignParticipationId });
       expect(poleEmploiSendings.length).to.equal(1);
       expect(poleEmploiSendings[0].responseCode).to.equal(responseCode.toString());
-      expect(poleEmploiSendings[0].type).to.equal('CAMPAIGN_PARTICIPATION_COMPLETION');
+      expect(poleEmploiSendings[0].type).to.equal('CAMPAIGN_PARTICIPATION_START');
     });
   });
 });
