+++ api/tests/prescription/campaign-participation/unit/domain/events/handle-pole-emploi-participation-finished_test.js
@@ -1,16 +1,14 @@
-import { CampaignParticipationStarted } from '../../../../lib/domain/events/CampaignParticipationStarted.js';
-import { _forTestOnly } from '../../../../lib/domain/events/index.js';
-import { PoleEmploiSending } from '../../../../lib/domain/models/PoleEmploiSending.js';
-import { PoleEmploiPayload } from '../../../../lib/infrastructure/externals/pole-emploi/PoleEmploiPayload.js';
-import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
-const { handlePoleEmploiParticipationStarted } = _forTestOnly.handlers;
-import * as httpErrorsHelper from '../../../../lib/infrastructure/http/errors-helper.js';
-import { httpAgent } from '../../../../lib/infrastructure/http/http-agent.js';
-import * as monitoringTools from '../../../../lib/infrastructure/monitoring-tools.js';
+import { AssessmentCompleted } from '../../../../../../lib/domain/events/AssessmentCompleted.js';
+import { _forTestOnly } from '../../../../../../src/prescription/campaign-participation/domain/events/index.js';
+import { PoleEmploiSending } from '../../../../../../src/prescription/campaign-participation/domain/models/PoleEmploiSending.js';
+import { PoleEmploiPayload } from '../../../../../../src/prescription/campaign-participation/infrastructure/externals/PoleEmploiPayload.js';
+import { catchErr, domainBuilder, expect, sinon } from '../../../../../test-helper.js';
+const { handlePoleEmploiParticipationFinished } = _forTestOnly.handlers;
 
-describe('Unit | Domain | Events | handle-pole-emploi-participation-started', function () {
+describe('Unit | Domain | Events | handle-pole-emploi-participation-finished', function () {
   let event, dependencies, expectedResults;
-  let campaignRepository,
+  let assessmentRepository,
+    campaignRepository,
     campaignParticipationRepository,
     organizationRepository,
     targetProfileRepository,
@@ -20,10 +18,7 @@ describe('Unit | Domain | Events | handle-pole-emploi-participation-started', fu
     authenticationMethodRepository;
 
   beforeEach(function () {
-    authenticationMethodRepository = {
-      findOneByUserIdAndIdentityProvider: sinon.stub(),
-      updateAuthenticationComplementByUserIdAndIdentityProvider: sinon.stub(),
-    };
+    assessmentRepository = { get: sinon.stub() };
     campaignRepository = { get: sinon.stub() };
     campaignParticipationRepository = { get: sinon.stub() };
     organizationRepository = { get: sinon.stub() };
@@ -31,16 +26,21 @@ describe('Unit | Domain | Events | handle-pole-emploi-participation-started', fu
     userRepository = { get: sinon.stub() };
     poleEmploiNotifier = { notify: sinon.stub() };
     poleEmploiSendingRepository = { create: sinon.stub() };
+    authenticationMethodRepository = {
+      findOneByUserIdAndIdentityProvider: sinon.stub(),
+      updateAuthenticationComplementByUserIdAndIdentityProvider: sinon.stub(),
+    };
 
     dependencies = {
-      authenticationMethodRepository,
+      assessmentRepository,
       campaignRepository,
       campaignParticipationRepository,
       organizationRepository,
-      poleEmploiSendingRepository,
       targetProfileRepository,
       userRepository,
       poleEmploiNotifier,
+      poleEmploiSendingRepository,
+      authenticationMethodRepository,
     };
 
     expectedResults = new PoleEmploiPayload({
@@ -59,12 +59,12 @@ describe('Unit | Domain | Events | handle-pole-emploi-participation-started', fu
         prenom: 'Jean',
       },
       test: {
-        etat: 2,
-        progression: 0,
+        etat: 3,
+        progression: 100,
         typeTest: 'DI',
         referenceExterne: 55667788,
         dateDebut: new Date('2020-01-02'),
-        dateProgression: null,
+        dateProgression: new Date('2020-01-03'),
         dateValidation: null,
         evaluation: null,
         uniteEvaluation: 'A',
@@ -77,20 +77,20 @@ describe('Unit | Domain | Events | handle-pole-emploi-participation-started', fu
     // given
     const event = 'not an event of the correct type';
     // when / then
-    const error = await catchErr(handlePoleEmploiParticipationStarted)({ event, ...dependencies });
+    const error = await catchErr(handlePoleEmploiParticipationFinished)({ event, ...dependencies });
 
     // then
     expect(error).not.to.be.null;
   });
 
-  context('#handlePoleEmploiParticipationStarted', function () {
-    let campaignParticipationId, campaignId, userId, organizationId;
+  context('#handlePoleEmploiParticipationFinished', function () {
+    let campaignId, userId, organizationId, assessmentId;
 
     beforeEach(function () {
-      campaignParticipationId = 55667788;
       campaignId = Symbol('campaignId');
       userId = Symbol('userId');
       organizationId = Symbol('organizationId');
+      assessmentId = Symbol('assessmentId');
     });
 
     context('when campaign is of type ASSESSMENT and organization is Pole Emploi', function () {
@@ -106,39 +106,38 @@ describe('Unit | Domain | Events | handle-pole-emploi-participation-started', fu
           organization: { id: organizationId },
         });
         const campaignParticipation = domainBuilder.buildCampaignParticipation({
-          id: campaignParticipationId,
+          id: 55667788,
           userId,
           campaign,
+          assessments: [domainBuilder.buildAssessment({ id: assessmentId })],
           createdAt: new Date('2020-01-02'),
         });
+        event = new AssessmentCompleted({ campaignParticipationId: campaignParticipation.id });
 
-        campaignParticipationRepository.get.withArgs(campaignParticipationId).resolves(campaignParticipation);
-        campaignRepository.get.withArgs(campaignId).resolves(campaign);
+        campaignParticipationRepository.get.withArgs(campaignParticipation.id).resolves(campaignParticipation);
+        assessmentRepository.get.withArgs(assessmentId).resolves(
+          domainBuilder.buildAssessment({
+            updatedAt: new Date('2020-01-03'),
+          }),
+        );
         organizationRepository.get.withArgs(organizationId).resolves({ isPoleEmploi: true });
         userRepository.get
           .withArgs(userId)
           .resolves(domainBuilder.buildUser({ id: userId, firstName: 'Jean', lastName: 'Bonneau' }));
+        campaignRepository.get.withArgs(campaignId).resolves(campaign);
         targetProfileRepository.get.withArgs('targetProfileId1').resolves({ name: 'Diagnostic initial' });
-
-        event = new CampaignParticipationStarted({ campaignParticipationId });
       });
 
       it('should notify pole emploi and create pole emploi sending accordingly', async function () {
         // given
+
         const expectedResponse = { isSuccessful: 'someValue', code: 'someCode' };
-        poleEmploiNotifier.notify
-          .withArgs(userId, expectedResults, {
-            authenticationMethodRepository,
-            httpAgent,
-            httpErrorsHelper,
-            monitoringTools,
-          })
-          .resolves(expectedResponse);
+        poleEmploiNotifier.notify.withArgs(userId, expectedResults).resolves(expectedResponse);
         const poleEmploiSending = Symbol('Pole emploi sending');
         sinon
-          .stub(PoleEmploiSending, 'buildForParticipationStarted')
+          .stub(PoleEmploiSending, 'buildForParticipationFinished')
           .withArgs({
-            campaignParticipationId,
+            campaignParticipationId: 55667788,
             payload: expectedResults.toString(),
             isSuccessful: expectedResponse.isSuccessful,
             responseCode: expectedResponse.code,
@@ -146,7 +145,7 @@ describe('Unit | Domain | Events | handle-pole-emploi-participation-started', fu
           .returns(poleEmploiSending);
 
         // when
-        await handlePoleEmploiParticipationStarted({
+        await handlePoleEmploiParticipationFinished({
           event,
           ...dependencies,
         });
@@ -164,22 +163,21 @@ describe('Unit | Domain | Events | handle-pole-emploi-participation-started', fu
           organization: { id: organizationId },
         });
         const campaignParticipation = domainBuilder.buildCampaignParticipation({
-          id: campaignParticipationId,
+          id: 55667788,
           userId,
           campaign,
           createdAt: new Date('2020-01-02'),
         });
+        event = new AssessmentCompleted({ campaignParticipationId: campaignParticipation.id });
 
-        campaignParticipationRepository.get.withArgs(campaignParticipationId).resolves(campaignParticipation);
-        campaignRepository.get.withArgs(campaignId).resolves(campaign);
+        campaignParticipationRepository.get.withArgs(campaignParticipation.id).resolves(campaignParticipation);
+        campaignRepository.get.withArgs(campaignId).resolves(domainBuilder.buildCampaign(campaign));
         organizationRepository.get.withArgs(organizationId).resolves({ isPoleEmploi: false });
-
-        event = new CampaignParticipationStarted({ campaignParticipationId });
       });
 
       it('should not notify to Pole Emploi', async function () {
         // when
-        await handlePoleEmploiParticipationStarted({
+        await handlePoleEmploiParticipationFinished({
           event,
           ...dependencies,
         });
@@ -197,26 +195,26 @@ describe('Unit | Domain | Events | handle-pole-emploi-participation-started', fu
           organization: { id: organizationId },
         });
         const campaignParticipation = domainBuilder.buildCampaignParticipation({
-          id: campaignParticipationId,
+          id: 55667788,
           userId,
           campaign,
           createdAt: new Date('2020-01-02'),
         });
+        event = new AssessmentCompleted({ campaignParticipationId: campaignParticipation.id });
 
-        campaignParticipationRepository.get.withArgs(campaignParticipationId).resolves(campaignParticipation);
-        campaignRepository.get.withArgs(campaignId).resolves(campaign);
+        campaignParticipationRepository.get.withArgs(campaignParticipation.id).resolves(campaignParticipation);
+        campaignRepository.get.withArgs(campaignId).resolves(domainBuilder.buildCampaign(campaign));
         organizationRepository.get.withArgs(organizationId).resolves({ isPoleEmploi: true });
-
-        event = new CampaignParticipationStarted({ campaignParticipationId });
       });
 
       it('should not notify to Pole Emploi', async function () {
         // when
-        await handlePoleEmploiParticipationStarted({
+        await handlePoleEmploiParticipationFinished({
           event,
           ...dependencies,
         });
 
+        // then
         sinon.assert.notCalled(poleEmploiNotifier.notify);
       });
     });
