+++ api/tests/prescription/campaign-participation/unit/domain/events/handle-pole-emploi-participation-started_test.js
@@ -1,17 +1,13 @@
-import { AssessmentCompleted } from '../../../../lib/domain/events/AssessmentCompleted.js';
-import { _forTestOnly } from '../../../../lib/domain/events/index.js';
-import { PoleEmploiSending } from '../../../../lib/domain/models/PoleEmploiSending.js';
-import { PoleEmploiPayload } from '../../../../lib/infrastructure/externals/pole-emploi/PoleEmploiPayload.js';
-import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
-const { handlePoleEmploiParticipationFinished } = _forTestOnly.handlers;
-import * as httpErrorsHelper from '../../../../lib/infrastructure/http/errors-helper.js';
-import { httpAgent } from '../../../../lib/infrastructure/http/http-agent.js';
-import * as monitoringTools from '../../../../lib/infrastructure/monitoring-tools.js';
+import { CampaignParticipationStarted } from '../../../../../../lib/domain/events/CampaignParticipationStarted.js';
+import { _forTestOnly } from '../../../../../../src/prescription/campaign-participation/domain/events/index.js';
+import { PoleEmploiSending } from '../../../../../../src/prescription/campaign-participation/domain/models/PoleEmploiSending.js';
+import { PoleEmploiPayload } from '../../../../../../src/prescription/campaign-participation/infrastructure/externals/PoleEmploiPayload.js';
+import { catchErr, domainBuilder, expect, sinon } from '../../../../../test-helper.js';
+const { handlePoleEmploiParticipationStarted } = _forTestOnly.handlers;
 
-describe('Unit | Domain | Events | handle-pole-emploi-participation-finished', function () {
+describe('Unit | Domain | Events | handle-pole-emploi-participation-started', function () {
   let event, dependencies, expectedResults;
-  let assessmentRepository,
-    campaignRepository,
+  let campaignRepository,
     campaignParticipationRepository,
     organizationRepository,
     targetProfileRepository,
@@ -21,7 +17,10 @@ describe('Unit | Domain | Events | handle-pole-emploi-participation-finished', f
     authenticationMethodRepository;
 
   beforeEach(function () {
-    assessmentRepository = { get: sinon.stub() };
+    authenticationMethodRepository = {
+      findOneByUserIdAndIdentityProvider: sinon.stub(),
+      updateAuthenticationComplementByUserIdAndIdentityProvider: sinon.stub(),
+    };
     campaignRepository = { get: sinon.stub() };
     campaignParticipationRepository = { get: sinon.stub() };
     organizationRepository = { get: sinon.stub() };
@@ -29,21 +28,16 @@ describe('Unit | Domain | Events | handle-pole-emploi-participation-finished', f
     userRepository = { get: sinon.stub() };
     poleEmploiNotifier = { notify: sinon.stub() };
     poleEmploiSendingRepository = { create: sinon.stub() };
-    authenticationMethodRepository = {
-      findOneByUserIdAndIdentityProvider: sinon.stub(),
-      updateAuthenticationComplementByUserIdAndIdentityProvider: sinon.stub(),
-    };
 
     dependencies = {
-      assessmentRepository,
+      authenticationMethodRepository,
       campaignRepository,
       campaignParticipationRepository,
       organizationRepository,
+      poleEmploiSendingRepository,
       targetProfileRepository,
       userRepository,
       poleEmploiNotifier,
-      poleEmploiSendingRepository,
-      authenticationMethodRepository,
     };
 
     expectedResults = new PoleEmploiPayload({
@@ -62,12 +56,12 @@ describe('Unit | Domain | Events | handle-pole-emploi-participation-finished', f
         prenom: 'Jean',
       },
       test: {
-        etat: 3,
-        progression: 100,
+        etat: 2,
+        progression: 0,
         typeTest: 'DI',
         referenceExterne: 55667788,
         dateDebut: new Date('2020-01-02'),
-        dateProgression: new Date('2020-01-03'),
+        dateProgression: null,
         dateValidation: null,
         evaluation: null,
         uniteEvaluation: 'A',
@@ -80,20 +74,20 @@ describe('Unit | Domain | Events | handle-pole-emploi-participation-finished', f
     // given
     const event = 'not an event of the correct type';
     // when / then
-    const error = await catchErr(handlePoleEmploiParticipationFinished)({ event, ...dependencies });
+    const error = await catchErr(handlePoleEmploiParticipationStarted)({ event, ...dependencies });
 
     // then
     expect(error).not.to.be.null;
   });
 
-  context('#handlePoleEmploiParticipationFinished', function () {
-    let campaignId, userId, organizationId, assessmentId;
+  context('#handlePoleEmploiParticipationStarted', function () {
+    let campaignParticipationId, campaignId, userId, organizationId;
 
     beforeEach(function () {
+      campaignParticipationId = 55667788;
       campaignId = Symbol('campaignId');
       userId = Symbol('userId');
       organizationId = Symbol('organizationId');
-      assessmentId = Symbol('assessmentId');
     });
 
     context('when campaign is of type ASSESSMENT and organization is Pole Emploi', function () {
@@ -109,45 +103,32 @@ describe('Unit | Domain | Events | handle-pole-emploi-participation-finished', f
           organization: { id: organizationId },
         });
         const campaignParticipation = domainBuilder.buildCampaignParticipation({
-          id: 55667788,
+          id: campaignParticipationId,
           userId,
           campaign,
-          assessments: [domainBuilder.buildAssessment({ id: assessmentId })],
           createdAt: new Date('2020-01-02'),
         });
-        event = new AssessmentCompleted({ campaignParticipationId: campaignParticipation.id });
 
-        campaignParticipationRepository.get.withArgs(campaignParticipation.id).resolves(campaignParticipation);
-        assessmentRepository.get.withArgs(assessmentId).resolves(
-          domainBuilder.buildAssessment({
-            updatedAt: new Date('2020-01-03'),
-          }),
-        );
+        campaignParticipationRepository.get.withArgs(campaignParticipationId).resolves(campaignParticipation);
+        campaignRepository.get.withArgs(campaignId).resolves(campaign);
         organizationRepository.get.withArgs(organizationId).resolves({ isPoleEmploi: true });
         userRepository.get
           .withArgs(userId)
           .resolves(domainBuilder.buildUser({ id: userId, firstName: 'Jean', lastName: 'Bonneau' }));
-        campaignRepository.get.withArgs(campaignId).resolves(campaign);
         targetProfileRepository.get.withArgs('targetProfileId1').resolves({ name: 'Diagnostic initial' });
+
+        event = new CampaignParticipationStarted({ campaignParticipationId });
       });
 
       it('should notify pole emploi and create pole emploi sending accordingly', async function () {
         // given
-
         const expectedResponse = { isSuccessful: 'someValue', code: 'someCode' };
-        poleEmploiNotifier.notify
-          .withArgs(userId, expectedResults, {
-            authenticationMethodRepository,
-            httpAgent,
-            httpErrorsHelper,
-            monitoringTools,
-          })
-          .resolves(expectedResponse);
+        poleEmploiNotifier.notify.withArgs(userId, expectedResults).resolves(expectedResponse);
         const poleEmploiSending = Symbol('Pole emploi sending');
         sinon
-          .stub(PoleEmploiSending, 'buildForParticipationFinished')
+          .stub(PoleEmploiSending, 'buildForParticipationStarted')
           .withArgs({
-            campaignParticipationId: 55667788,
+            campaignParticipationId,
             payload: expectedResults.toString(),
             isSuccessful: expectedResponse.isSuccessful,
             responseCode: expectedResponse.code,
@@ -155,7 +136,7 @@ describe('Unit | Domain | Events | handle-pole-emploi-participation-finished', f
           .returns(poleEmploiSending);
 
         // when
-        await handlePoleEmploiParticipationFinished({
+        await handlePoleEmploiParticipationStarted({
           event,
           ...dependencies,
         });
@@ -173,21 +154,22 @@ describe('Unit | Domain | Events | handle-pole-emploi-participation-finished', f
           organization: { id: organizationId },
         });
         const campaignParticipation = domainBuilder.buildCampaignParticipation({
-          id: 55667788,
+          id: campaignParticipationId,
           userId,
           campaign,
           createdAt: new Date('2020-01-02'),
         });
-        event = new AssessmentCompleted({ campaignParticipationId: campaignParticipation.id });
 
-        campaignParticipationRepository.get.withArgs(campaignParticipation.id).resolves(campaignParticipation);
-        campaignRepository.get.withArgs(campaignId).resolves(domainBuilder.buildCampaign(campaign));
+        campaignParticipationRepository.get.withArgs(campaignParticipationId).resolves(campaignParticipation);
+        campaignRepository.get.withArgs(campaignId).resolves(campaign);
         organizationRepository.get.withArgs(organizationId).resolves({ isPoleEmploi: false });
+
+        event = new CampaignParticipationStarted({ campaignParticipationId });
       });
 
       it('should not notify to Pole Emploi', async function () {
         // when
-        await handlePoleEmploiParticipationFinished({
+        await handlePoleEmploiParticipationStarted({
           event,
           ...dependencies,
         });
@@ -205,26 +187,26 @@ describe('Unit | Domain | Events | handle-pole-emploi-participation-finished', f
           organization: { id: organizationId },
         });
         const campaignParticipation = domainBuilder.buildCampaignParticipation({
-          id: 55667788,
+          id: campaignParticipationId,
           userId,
           campaign,
           createdAt: new Date('2020-01-02'),
         });
-        event = new AssessmentCompleted({ campaignParticipationId: campaignParticipation.id });
 
-        campaignParticipationRepository.get.withArgs(campaignParticipation.id).resolves(campaignParticipation);
-        campaignRepository.get.withArgs(campaignId).resolves(domainBuilder.buildCampaign(campaign));
+        campaignParticipationRepository.get.withArgs(campaignParticipationId).resolves(campaignParticipation);
+        campaignRepository.get.withArgs(campaignId).resolves(campaign);
         organizationRepository.get.withArgs(organizationId).resolves({ isPoleEmploi: true });
+
+        event = new CampaignParticipationStarted({ campaignParticipationId });
       });
 
       it('should not notify to Pole Emploi', async function () {
         // when
-        await handlePoleEmploiParticipationFinished({
+        await handlePoleEmploiParticipationStarted({
           event,
           ...dependencies,
         });
 
-        // then
         sinon.assert.notCalled(poleEmploiNotifier.notify);
       });
     });
