+++ api/tests/unit/domain/events/event-choreography-certification-rescoring-after-jury_test.js
@@ -1,11 +1,12 @@
 import { CertificationJuryDone } from '../../../../lib/domain/events/CertificationJuryDone.js';
+import { _forTestOnly } from '../../../../lib/domain/events/index.js';
 import { expect } from '../../../test-helper.js';
 import { buildEventDispatcherAndHandlersForTest } from '../../../tooling/events/event-dispatcher-builder.js';
 
 describe('Event Choreography | CertificationJuryDone', function () {
   it('Should trigger the certification rescoring', async function () {
     // given
-    const { handlerStubs, eventDispatcher } = buildEventDispatcherAndHandlersForTest();
+    const { handlerStubs, eventDispatcher } = buildEventDispatcherAndHandlersForTest(_forTestOnly);
     const event = new CertificationJuryDone({});
 
     // when
