+++ api/tests/unit/domain/events/event-choreography-pole-emploi-participation-finished_test.js
@@ -1,11 +1,12 @@
 import { AssessmentCompleted } from '../../../../lib/domain/events/AssessmentCompleted.js';
+import { _forTestOnly } from '../../../../src/prescription/campaign-participation/domain/events/index.js';
 import { expect } from '../../../test-helper.js';
 import { buildEventDispatcherAndHandlersForTest } from '../../../tooling/events/event-dispatcher-builder.js';
 
 describe('Event Choreography | Pole Emploi Participation Finished', function () {
   it('Should trigger Pole Emploi participation finished handler on AssessmentCompleted event', async function () {
     // given
-    const { handlerStubs, eventDispatcher } = buildEventDispatcherAndHandlersForTest();
+    const { handlerStubs, eventDispatcher } = buildEventDispatcherAndHandlersForTest(_forTestOnly);
     const event = new AssessmentCompleted();
     const domainTransaction = Symbol('a transaction');
