+++ api/tests/unit/domain/usecases/send-shared-participation-results-to-pole-emploi_test.js
@@ -1,9 +1,9 @@
-import { PoleEmploiSending } from '../../../../lib/domain/models/PoleEmploiSending.js';
-import { sendSharedParticipationResultsToPoleEmploi } from '../../../../lib/domain/usecases/send-shared-participation-results-to-pole-emploi.js';
-import { PoleEmploiPayload } from '../../../../lib/infrastructure/externals/pole-emploi/PoleEmploiPayload.js';
 import * as httpErrorsHelper from '../../../../lib/infrastructure/http/errors-helper.js';
 import { httpAgent } from '../../../../lib/infrastructure/http/http-agent.js';
 import * as monitoringTools from '../../../../lib/infrastructure/monitoring-tools.js';
+import { PoleEmploiSending } from '../../../../src/prescription/campaign-participation/domain/models/PoleEmploiSending.js';
+import { sendSharedParticipationResultsToPoleEmploi } from '../../../../src/prescription/campaign-participation/domain/usecases/send-shared-participation-results-to-pole-emploi.js';
+import { PoleEmploiPayload } from '../../../../src/prescription/campaign-participation/infrastructure/externals/PoleEmploiPayload.js';
 import { domainBuilder, expect, sinon } from '../../../test-helper.js';
 
 describe('Unit | Domain | UseCase | send-shared-participation-results-to-pole-emploi', function () {
