+++ api/tests/unit/infrastructure/externals/pole-emploi/pole-emploi-notifier_test.js
@@ -4,9 +4,9 @@ import dayjs from 'dayjs';
 
 import { config as settings } from '../../../../../lib/config.js';
 import { UnexpectedUserAccountError } from '../../../../../lib/domain/errors.js';
-import { notify } from '../../../../../lib/infrastructure/externals/pole-emploi/pole-emploi-notifier.js';
 import * as OidcIdentityProviders from '../../../../../src/identity-access-management/domain/constants/oidc-identity-providers.js';
 import { AuthenticationMethod } from '../../../../../src/identity-access-management/domain/models/AuthenticationMethod.js';
+import { notify } from '../../../../../src/prescription/campaign-participation/infrastructure/externals/pole-emploi-notifier.js';
 import { catchErr, expect, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Infrastructure | Externals/Pole-Emploi | pole-emploi-notifier', function () {
